+++ elements/user-picker/docs/0-intro.tsx
@@ -16,7 +16,7 @@ export default md`
   #### DISCLAIMER:
 
   Please consider using the smart-user-picker over user-picker if possible. This encourages consistency and monitoring.
-  If you cannot use smart-user-picker, please reach out to the our slack channels, #smrt-user-picker or #smrt-experiences,
+  If you cannot use smart-user-picker, please reach out to the our slack channels, #help-smart-experiences,
   so that we can work with you.
 
   ## Usage