+++ elements/user-picker/src/components/smart-user-picker/index.ts
@@ -1,3 +1,7 @@
+/**
+ * Please use AK/SUP instead
+ * @deprecated
+ */
 export { default as SmartUserPicker, hydrateDefaultValues } from './components';
 export { getUserRecommendations } from './service';
 export type {